Add view and integration tests for withdrawl
This also adds some custom styling for the dividers so they look more
like subheadings
