+++ assets/sass/application.sass
@@ -14,6 +14,7 @@ $path: "/assets/images/"
 @import './components/bed-search-inputs'
 @import './components/filter'
 @import './components/calendar'
+@import './components/withdrawal-reasons'
 
 @import './pages/assessments-index'
 @import './pages/applications-pages-attach-document'
