+++ integration_tests/pages/apply/withdrawApplicationPage.ts
@@ -4,11 +4,11 @@ import Page from '../page'
 
 export default class WithdrawApplicationPage extends Page {
   constructor() {
-    super('Do you want to withdraw this application?')
+    super('Why is this application being withdrawn?')
     this.checkForBackButton(paths.applications.index.pattern)
   }
 
-  clickYes() {
-    this.checkRadioByNameAndValue('confirmWithdrawal', 'yes')
+  completeForm() {
+    this.checkRadioByNameAndValue('reason', 'alternative_identified_placement_no_longer_required')
   }
 }
