--- server/views/applications/withdraw.njk
@@ -1,61 +1,0 @@
-{% from "govuk/components/button/macro.njk" import govukButton %}
-{% from "../partials/showErrorSummary.njk" import showErrorSummary %}
-{% from "govuk/components/radios/macro.njk" import govukRadios %}
-{% from "govuk/components/back-link/macro.njk" import govukBackLink %}
-
-{% extends "../partials/layout.njk" %}
-
-{% set pageTitle = applicationName + " - " + pageHeading  %}
-{% set mainClasses = "app-container govuk-body" %}
-
-{% block beforeContent %}
-    {{ govukBackLink({
-      text: "Back",
-      href: paths.applications.index({})
-    }) }}
-{% endblock %}
-
-{% block content %}
-
-    <div class="govuk-grid-row">
-        <div class="govuk-grid-column-two-thirds">
-            <form action="{{ paths.applications.withdraw.create({id: applicationId}) }}" method="post">
-                <input type="hidden" name="_csrf" value="{{ csrfToken }}"/>
-
-                {{ showErrorSummary(errorSummary) }}
-
-                {{ govukRadios({
-                        idPrefix: "confirmWithdrawal",
-                        name: "confirmWithdrawal",
-                        fieldset: {
-                            legend: {
-                                text: "Do you want to withdraw this application?",
-                                classes: "govuk-fieldset__legend--l",
-                                isPageHeading: true
-                            }
-                        },
-                        hint: {
-                            text: 'Withdrawing an application will remove the application data from the AP service. It cannot be undone.'
-                        },
-                        items: [
-                            {
-                            value: "yes",
-                            text: "Yes"
-                            },
-                            {
-                            value: "no",
-                            text: "No"
-                            }
-                        ],
-                        errorMessage: errors.confirmWithdrawal
-                     }) }}
-
-                {{ govukButton({
-                    name: 'submit',
-                    text: "Continue"
-                }) }}
-
-            </form>
-        </div>
-    </div>
-{% endblock %}
